packages/my-calculator-app/my_calculator_app-0.1.1.tar.gz/my_calculator_app-0.1.1/calculator/calculator.py
class Calculator:

    def add(self, a: float, b: float) -> float:
       """Return the sum of a and b."""
       return a + b

    def subtract(self, a: float, b: float) -> float:
        """Return the difference of a and b."""
        return a - b

    def multiply(self, a: float, b: float) -> float:
        """Return the product of a and b."""
        return a * b

    def divide(self, a: float, b: float) -> float:
        """Return the quotient of a and b. Raise ValueError if b is zero."""
        if b == 0:
            raise ValueError("Division by zero is not allowed.")
        return a / b
